refactor(api): make the safe wrapper Rust-native and reorganize the docs
- replace the `sys` type aliases with real enums (PixelFormat, ImageLabel,
  ColorGamut/Transfer/Range, Codec, Preset) plus From/TryFrom conversions
- make RawImage/CompressedImage own their bytes instead of borrowing them;
  upstream deep-copies on registration, so this is sound and drops every
  lifetime from the public API
- turn Error into an enum (InvalidParameter/InvalidOperation/...) with
  Display, std::error::Error, code() and detail()
- add Decoder::info()/decode_as()/rows()/decoded_gainmap(), ColorAspects
  builders, u8 quality with range checks, and generic set_exif_data; merge the
  three set_raw_image variants, keeping the zero-copy set_decoded_image
- split types.rs into image.rs + gainmap.rs, document the six public modules,
  deny missing docs and keep `cargo doc` warning-free
- port the CLI, the example, the integration tests and README to the new API

+++ ultrahdr-bake/src/cli.rs
@@ -62,18 +62,18 @@ pub struct BakeArgs {
     #[arg(
         long = "base-q",
         default_value_t = 95,
-        value_parser = clap::value_parser!(i32).range(1..=100)
+        value_parser = clap::value_parser!(u8).range(1..=100)
     )]
-    pub base_quality: i32,
+    pub base_quality: u8,
 
     /// JPEG quality for the gain map (1-100)
     #[arg(
         long = "gm-q",
         alias = "gainmap-q",
         default_value_t = 100,
-        value_parser = clap::value_parser!(i32).range(1..=100)
+        value_parser = clap::value_parser!(u8).range(1..=100)
     )]
-    pub gainmap_quality: i32,
+    pub gainmap_quality: u8,
 
     /// Gain map scale factor
     #[arg(
